+++ NEWS.md
@@ -1,3 +1,7 @@
+# getRad (development version)
+
+* New function `weather_radars()` retrieves metadata for OPERA weather radars (#15).
+
 # getRad 0.1.0
 
 * Initial package development